refactor: remove comments

+++ src/main/java/com/addventure/AddVenture/model/Etiqueta.java
@@ -4,12 +4,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
+import lombok.*;
 
 @Entity
 @Table(name = "Etiqueta")
@@ -19,18 +14,17 @@ import lombok.ToString;
 @Builder
 public class Etiqueta {
 
-    //Declarar los campos o atributos
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_etiqueta")
-    private Integer idEtiqueta;
+    private Long idEtiqueta;
 
     @Column(name = "nombre_etiqueta", nullable = false, length = 50, unique = true)
     private String nombreEtiqueta;
 
-    //Declarar las relaciones
     @EqualsAndHashCode.Exclude
     @ToString.Exclude
     @ManyToMany(mappedBy = "etiquetas")
     private Set<GrupoViaje> grupos = new HashSet<>();
+    
 }
